refactored Sensoe->getValue(). added float return typehint

+++ Tests/UnitTests/Sensors/Sensors/System/MemoryUsedTest.php
@@ -35,7 +35,7 @@ class MemoryUsedTest extends TestCase
     {
         $sensor = new SensorVO();
         $actual = $this->subject->getValue($sensor);
-        $this->assertInternalType('int', $actual);
+        $this->assertInternalType('float', $actual);
     }
 
     public function testGetDefinition()
